# docs/specs/swagger.json
{"openapi":"3.0.1","info":{"title":"'Practica Parcial 1'","description":"'LCIV Practica de parcial con uso de Docker y Docker Compose'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8080","description":"'LCIV Practica de parcial con uso de Docker y Docker Compose'"}],"paths":{"/user":{"post":{"tags":["user-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserResponseDTO"}}}}}}},"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/location":{"get":{"tags":["user-controller"],"operationId":"getLocations","parameters":[{"name":"client_id","in":"header","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"client_secret","in":"header","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Location"}}}}}}}},"/location/{id}":{"get":{"tags":["user-controller"],"operationId":"getWeatherByLocation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"datetime","in":"query","required":true,"schema":{"type":"string"}},{"name":"client_id","in":"header","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"client_secret","in":"header","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WeatherResponseDTO"}}}}}}}},"components":{"schemas":{"UserRequestDTO":{"type":"object","properties":{"email":{"type":"string"},"temperature_unit":{"type":"string"}}},"UserResponseDTO":{"type":"object","properties":{"client_id":{"type":"integer","format":"int64"},"secret":{"type":"string","format":"uuid"}}},"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}},"Location":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"latitude":{"type":"string"},"longitude":{"type":"string"}}},"AirQuality":{"type":"object","properties":{"index":{"type":"integer","format":"int64"},"description":{"type":"string"}}},"Cloudiness":{"type":"object","properties":{"index":{"type":"integer","format":"int64"},"description":{"type":"string"}}},"Temperature":{"type":"object","properties":{"value":{"type":"integer","format":"int64"},"unit":{"type":"string"}}},"WeatherResponseDTO":{"type":"object","properties":{"location":{"$ref":"#/components/schemas/Location"},"temperature":{"$ref":"#/components/schemas/Temperature"},"wind":{"type":"string"},"airQuality":{"$ref":"#/components/schemas/AirQuality"},"cloudiness":{"$ref":"#/components/schemas/Cloudiness"}}}}}}